If applied, this commit will add the tikka page
This applies the final part of the Project: Recipes adding the
Chicken Tikka Burger page to the landing page, 'index.html'.
The hyperlink references (href's) will take you to the other pages
in this project under the names of Chicken Tikka Burger, which was
added in this commit, and the lasagna and chocolate chip cookies
pages in the previous commit.